rename database to documentstore
