+++ .gitignore
@@ -138,7 +138,7 @@ dmypy.json
 .idea
 
 # haystack files
-haystack/database/qa.db
+haystack/documentstore/qa.db
 data
 mlruns
 src
