+++ haystack/documentstore/base.py
