+++ haystack/reader/farm.py
@@ -19,7 +19,7 @@ from farm.utils import set_all_seeds, initialize_device_settings
 from scipy.special import expit
 import shutil
 
-from haystack.database.base import Document, BaseDocumentStore
+from haystack.documentstore.base import Document, BaseDocumentStore
 from haystack.reader.base import BaseReader
 logger = logging.getLogger(__name__)
