+++ haystack/retriever/dense.py
@@ -6,8 +6,8 @@ from pathlib import Path
 
 from farm.infer import Inferencer
 
-from haystack.database.base import Document, BaseDocumentStore
-from haystack.database.elasticsearch import ElasticsearchDocumentStore
+from haystack.documentstore.base import Document, BaseDocumentStore
+from haystack.documentstore.elasticsearch import ElasticsearchDocumentStore
 from haystack.retriever.base import BaseRetriever
 from haystack.retriever.sparse import logger
