+++ rest_api/controller/feedback.py
@@ -3,7 +3,7 @@ from typing import Optional
 from fastapi import APIRouter
 from pydantic import BaseModel, Field
 
-from haystack.database.elasticsearch import ElasticsearchDocumentStore
+from haystack.documentstore.elasticsearch import ElasticsearchDocumentStore
 from rest_api.config import (
     DB_HOST,
     DB_PORT,
