+++ test/test_dummy_retriever.py
@@ -1,4 +1,4 @@
-from haystack.database.base import Document
+from haystack.documentstore.base import Document
 import pytest
