+++ tutorials/Tutorial4_FAQ_style_QA.py
@@ -1,5 +1,5 @@
 from haystack import Finder
-from haystack.database.elasticsearch import ElasticsearchDocumentStore
+from haystack.documentstore.elasticsearch import ElasticsearchDocumentStore
 
 from haystack.retriever.dense import EmbeddingRetriever
 from haystack.utils import print_answers
